Перенос Моделей и Моделей-Представлений в проект предметной области
