+++ Common/WebStore.Domain/Models/Employee.cs
@@ -3,14 +3,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebStore.Models
+namespace WebStore.Domain.Models
 {
     public class Employee
     {
         public int id { get; set; }
-        public String FirstName { get; set; }
-        public String SurName { get; set; }
-        public String Patronymic { get; set; }
+        public string FirstName { get; set; }
+        public string SurName { get; set; }
+        public string Patronymic { get; set; }
         public int Age { get; set; }
