+++ Common/WebStore.Domain/ViewModels/BrandViewModel.cs
@@ -4,16 +4,12 @@ using System.Linq;
 using System.Threading.Tasks;
 using WebStore.Domain.Entities.Base.Interfaces;
 
-namespace WebStore.ViewModels
+namespace WebStore.Domain.ViewModels
 {
-    public class ProductViewModel : INamedEntity, IOrderEntity
+    public class BrandViewModel : INamedEntity, IOrderEntity
     {
-        
         public int Id { get; set; }
         public string Name { get; set; }
         public int Order { get; set; }
-        public string ImageUrl { get; set; }
-        public decimal Price { get; set; }
-        public string Brand { get; set; }
     }
 }
