+++ Common/WebStore.Domain/ViewModels/CartOrderViewModel.cs
@@ -1,6 +1,7 @@
-﻿using WebStore.ViewModels.Orders;
+﻿using WebStore.Domain.ViewModels;
+using WebStore.Domain.ViewModels.Orders;
 
-namespace WebStore.ViewModels
+namespace WebStore.Domain.ViewModels
 {
     public class CartOrderViewModel
     {
