+++ Common/WebStore.Domain/ViewModels/Identity/RegisterUserViewModel.cs
@@ -1,13 +1,12 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebStore.ViewModels.Identity
+namespace WebStore.Domain.ViewModels.Identity
 {
-    public class LoginViewModel
+    public class RegisterUserViewModel
     {
         [Required]
         [MaxLength(250)]
@@ -19,12 +18,11 @@ namespace WebStore.ViewModels.Identity
         [Display(Name = "Пароль")]
         public string Password { get; set; }
 
-        [Display(Name = "Запомни меня")]
+        [Required]
+        [DataType(DataType.Password)]
+        [Display(Name = "Повторите пароль")]
         [Compare(nameof(Password))]
-        public bool RememberMy { get; set; }
-
-        [HiddenInput(DisplayValue = false)]
-        public string ReturnUrl { get; set; }
+        public string ConfirmPassword { get; set; }
 
     }
 }
