+++ Common/WebStore.Domain/ViewModels/ProductViewModel.cs
@@ -4,12 +4,16 @@ using System.Linq;
 using System.Threading.Tasks;
 using WebStore.Domain.Entities.Base.Interfaces;
 
-namespace WebStore.ViewModels
+namespace WebStore.Domain.ViewModels
 {
-    public class BrandViewModel : INamedEntity, IOrderEntity
+    public class ProductViewModel : INamedEntity, IOrderEntity
     {
+        
         public int Id { get; set; }
         public string Name { get; set; }
         public int Order { get; set; }
+        public string ImageUrl { get; set; }
+        public decimal Price { get; set; }
+        public string Brand { get; set; }
     }
 }
