+++ Common/WebStore.Domain/ViewModels/SectionViewModel.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using WebStore.Domain.Entities.Base.Interfaces;
 
-namespace WebStore.ViewModels
+namespace WebStore.Domain.ViewModels
 {
     public class SectionViewModel : INamedEntity, IOrderEntity
     {
