+++ UI/WebStore/Areas/Admin/Views/_ViewImports.cshtml
@@ -1,6 +1,6 @@
-﻿@using WebStore.Models
-@using WebStore.ViewModels
-@using WebStore.ViewModels.Identity
+﻿@using WebStore.Domain.Models
+@using WebStore.Domain.ViewModels
+@using WebStore.Domain.ViewModels.Identity
 @using WebStore.Domain.Entities
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
