+++ UI/WebStore/Controllers/HomeController.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using WebStore.Models;
+using WebStore.Domain.Models;
 
 namespace WebStore.Controllers
 {
