+++ UI/WebStore/Infrastructure/Interfaces/IOrderService.cs
@@ -1,8 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using WebStore.Domain.Entities.Orders;
-using WebStore.ViewModels;
-using WebStore.ViewModels.Orders;
+using WebStore.Domain.ViewModels;
+using WebStore.Domain.ViewModels.Orders;
+using WebStore.Domain.ViewModels.Orders;
 
 namespace WebStore.Infrastructure.Interfaces
 {
@@ -13,5 +14,6 @@ namespace WebStore.Infrastructure.Interfaces
         Order GetOrderById(int id);
 
         Task<Order> CreateOrderAsync(string UserName, CartViewModel Cart, OrderViewModel OrderModel);
+        Task CreateOrderAsync(string name, CartViewModel cartViewModel, OrderViewModel model);
     }
 }
