+++ UI/WebStore/Infrastructure/Mapping/EmployeeMapping.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using WebStore.Domain.Entities;
-using WebStore.Models;
-using WebStore.ViewModels;
+using WebStore.Domain.Models;
+using WebStore.Domain.ViewModels;
 
 namespace WebStore.Infrastructure.Mapping
 {
