+++ UI/WebStore/Infrastructure/Services/InSQL/SqlOrderService.cs
@@ -7,9 +7,9 @@ using Microsoft.EntityFrameworkCore;
 using WebStore.DAL.Context;
 using WebStore.Domain.Entities.Identity;
 using WebStore.Domain.Entities.Orders;
+using WebStore.Domain.ViewModels;
 using WebStore.Infrastructure.Interfaces;
-using WebStore.ViewModels;
-using WebStore.ViewModels.Orders;
+using WebStore.Domain.ViewModels.Orders;
 
 namespace WebStore.Infrastructure.Services.InSQL
 {
