+++ UI/WebStore/Views/Cart/Partial/_CheckOut.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebStore.ViewModels.Orders.OrderViewModel
+﻿@model WebStore.Domain.ViewModels.Orders.OrderViewModel
 
 <form method="post" asp-area="" asp-controller="Cart" asp-action="CheckOut" asp-antiforgery="true">
     <div class="row">
